+++ frontend/frontend/src/routes/index.jsx
@@ -1,14 +1,14 @@
 import { createBrowserRouter } from "react-router-dom";
-import LoginComponent from "../components/Login/LoginComponent";
-import ClientDashboard from "../components/Client/ClientDashboard";
-import VisiteurLayout from "../layouts/VisiteurLayout";
-import ClientDashboardLayout from "../layouts/ClientDashboardLayout";
-import AdminDashboardLayout from "../layouts/AdminDashboardLayout";
-import ClubDashboardLayout from "../layouts/ClubDashboardLayout";
-import AdminCalendrier from "../components/Admin/AdminCalendrier";
-import AdminDashboard from "../components/Admin/AdminDashboard";
-import ClubDashboard from "../components/Club/ClubDashboard";
-import AdminClub from "../components/Admin/AdminClub";
+import ClubDashboardLayout from "../club/layout/ClubDashboardLayout";
+import AdminDashboardLayout from "../admin/layout/AdminDashboardLayout";
+import VisiteurLayout from "../loginLayout/VisiteurLayout";
+import ClientDashboardLayout from "../client/layout/ClientDashboardLayout";
+import AdminDashboard from "../admin/components/AdminDashboard";
+import AdminCalendrier from "../admin/components/AdminCalendrier";
+import AdminClub from "../admin/components/AdminClub";
+import LoginComponent from "../loginLayout/LoginComponent";
+import ClientDashboard from "../client/components/ClientDashboard";
+import ClubDashboard from "../club/components/ClubDashboard";
 export const LOGIN_ROUTE ='/';
 export const CLIENT_DASHBOARD_ROUTE ='/client';
 export const ADMIN_DASHBOARD_ROUTE ='/admin';
